+++ app/apis/auth/service/auth.service.factory.ts
@@ -1,8 +1,8 @@
 import { AuthProvider } from "../../../types/enums/providers";
 import { IAuthService } from "../../../types/interfaces/auth.service.interface";
 import BaseServiceFactory from "../../base.service.factory";
-import FirebaseAuthService from "./providers/FirebaseAuthService";
-import MockAuthService from "./providers/MockAuthService";
+import FirebaseAuthService from "./providers/firebase.auth.service";
+import MockAuthService from "./providers/mock.auth.service";
 
 export default class AuthServiceFactory extends BaseServiceFactory<IAuthService, AuthProvider>{
     protected createService(serverPrivateKey:string, provider?: AuthProvider): IAuthService {
