+++ pages/api/v1/auth/login/firebase/custom-token.ts
@@ -1,7 +1,7 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import type { NextApiRequest, NextApiResponse } from "next";
 import { authServiceFactory } from "../../../../../../app/apis/services";
-import FirebaseAuthService from "../../../../../../app/apis/auth/service/providers/FirebaseAuthService";
+import FirebaseAuthService from "../../../../../../app/apis/auth/service/providers/firebase.auth.service";
 import { SessionDto } from "../../../../../../app/types/dtos/auth.dtos";
 import { ServerMessageDto } from "../../../../../../app/types/dtos/server-message.dtos";
 import { AuthProvider } from "../../../../../../app/types/enums/providers";
